% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleTCGA}
\alias{exampleTCGA}
\title{TCGA-LAML example data}
\format{
An object of class \code{data.frame} with 60483 rows and 21 columns.
}
\usage{
data(exampleTCGA)
}
\description{
TCGA-LAML example data
}
\keyword{datasets}
