% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleTCGAmeta}
\alias{exampleTCGAmeta}
\title{TCGA-LAML example data meta file}
\format{
An object of class \code{data.frame} with 9 rows and 35 columns.
}
\usage{
data(exampleTCGAmeta)
}
\description{
TCGA-LAML example data meta file
}
\keyword{datasets}
