% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minRes}
\alias{minRes}
\title{A minimal seAMLess result list object}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(minRes)
}
\description{
A minimal seAMLess result list object
}
\keyword{datasets}
