% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRef}
\alias{scRef}
\title{Single cell obj for deconvolution}
\format{
An object of class \code{ExpressionSet} with 439 rows and 69101 columns.
}
\usage{
data(scRef)
}
\description{
Single cell obj for deconvolution
}
\examples{
data(scRef)
}
\keyword{datasets}
