% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ternaryPlot}
\alias{ternaryPlot}
\title{Given the immune compositions (ICs) of bulk-RNA samples, this function creates
a ternary plot similar to ALOT tube from EuroFlow analysis and Figure 1E of our paper.}
\usage{
ternaryPlot(res)
}
\arguments{
\item{res}{seAMLess object.}
}
\value{
ggplot2 object
}
\description{
Given the immune compositions (ICs) of bulk-RNA samples, this function creates
a ternary plot similar to ALOT tube from EuroFlow analysis and Figure 1E of our paper.
}
\examples{
library(seAMLess)

data(minRes)
ternaryPlot(minRes)

}
