% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venoModel}
\alias{venoModel}
\title{Trained RF model on Venetoclax Resistance}
\format{
An object of class \code{randomForest} of length 17.
}
\usage{
data(venoModel)
}
\description{
Trained RF model on Venetoclax Resistance
}
\keyword{datasets}
