% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleMat}
\alias{wrangleMat}
\title{removes ERCC peaks and duplicated genes}
\usage{
wrangleMat(mat)
}
\arguments{
\item{mat}{pre-filters and orders bulk rna-seq data}
}
\description{
removes ERCC peaks and duplicated genes
}
