% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli}
\alias{cli}
\title{Command Line Interface for seAMLess}
\usage{
cli()
}
\value{
List of validated command line options
}
\description{
Provides a command line interface to run seAMLess deconvolution analysis
on bulk RNA-seq data.
}
