% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleTCGAmeta}
\alias{exampleTCGAmeta}
\title{TCGA-LAML example data meta file downloaded from GDC}
\format{
An object of class \code{data.frame} with 20 rows and 34 columns.
}
\usage{
data(exampleTCGAmeta)
}
\description{
TCGA-LAML example data meta file downloaded from GDC
}
\keyword{datasets}
