% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seAMLess.R
\name{seAMLess}
\alias{seAMLess}
\title{Given the count matrices of bulk-RNA samples, this function deconvolutes
each sample into its cell types using a healthy BM reference, and
calculates the sample's in vitro resistance to Venetoclax.}
\usage{
seAMLess(mat, verbose = TRUE)
}
\arguments{
\item{mat}{count matrix (genes by 1+samples).}

\item{verbose}{prints detailed messages}
}
\value{
List of deconvoluted cell type percentages and predicted drug resistances
}
\description{
Given the count matrices of bulk-RNA samples, this function deconvolutes
each sample into its cell types using a healthy BM reference, and
calculates the sample's in vitro resistance to Venetoclax.
}
